# controllers/book_controller.go
package controllers

import (
	"github.com/gofiber/fiber/v2"
	"github.com/zikrykr/backend-test-desent/infrastructure"
	"github.com/zikrykr/backend-test-desent/model"
	"github.com/zikrykr/backend-test-desent/services"
	"github.com/zikrykr/backend-test-desent/utils"
)

type BookController struct {
	BookService services.BookServiceInterface
	Logger      *infrastructure.Logger
}

func NewBookController(
	bookService services.BookServiceInterface,
	logger *infrastructure.Logger,
) BookControllerInterface {
	return &BookController{
		BookService: bookService,
		Logger:      logger,
	}
}

// Create returns a simple create response.
// @Summary Create
// @Description create a new book.
// @Tags book
// @Accept json
// @Produce json
// @Param body body model.CreateBookRequest true "Request body"
// @Success 201 {object} model.Book
// @Router /api/books [post]
func (c *BookController) Create(ctx *fiber.Ctx) error {
	var req model.CreateBookRequest
	if err := ctx.BodyParser(&req); err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusBadRequest, "failed to parse request body to valid JSON", err)
	}

	bookResult, err := c.BookService.Create(&req)
	if err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusInternalServerError, "failed to create book", err)
	}

	return utils.SuccessPlainResponse(ctx, c.Logger, fiber.StatusCreated, bookResult)
}

// FindByID returns a simple find by id response.
// @Summary FindByID
// @Description find a book by id.
// @Tags book
// @Accept */*
// @Produce json
// @Param id path string true "Book ID"
// @Success 200 {object} model.Book
// @Router /api/books/{id} [get]
func (c *BookController) FindByID(ctx *fiber.Ctx) error {
	id := ctx.Params("id")
	if id == "" {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusBadRequest, "book id cannot be empty", nil)
	}

	bookResult, err := c.BookService.FindByID(id)
	if err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusInternalServerError, "failed to find book", err)
	}

	return utils.SuccessPlainResponse(ctx, c.Logger, fiber.StatusOK, bookResult)
}

// FindAll returns a simple find all response.
// @Summary FindAll
// @Description find all books.
// @Tags book
// @Accept */*
// @Produce json
// @Success 200 {object} []model.Book
// @Router /api/books [get]
func (c *BookController) FindAll(ctx *fiber.Ctx) error {
	bookResult, err := c.BookService.FindAll()
	if err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusInternalServerError, "failed to find books", err)
	}

	return utils.SuccessPlainResponse(ctx, c.Logger, fiber.StatusOK, bookResult)
}

// Update returns a simple update response.
// @Summary Update
// @Description update a book.
// @Tags book
// @Accept json
// @Produce json
// @Param id path string true "Book ID"
// @Param body body model.UpdateBookRequest true "Request body"
// @Success 200 {object} model.Book
// @Router /api/books/{id} [put]
func (c *BookController) Update(ctx *fiber.Ctx) error {
	id := ctx.Params("id")
	if id == "" {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusBadRequest, "book id cannot be empty", nil)
	}

	var req model.UpdateBookRequest
	if err := ctx.BodyParser(&req); err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusBadRequest, "failed to parse request body to valid JSON", err)
	}

	bookResult, err := c.BookService.Update(id, &req)
	if err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusInternalServerError, "failed to update book", err)
	}

	return utils.SuccessPlainResponse(ctx, c.Logger, fiber.StatusOK, bookResult)
}

// Delete returns a simple delete response.
// @Summary Delete
// @Description delete a book.
// @Tags book
// @Accept */*
// @Produce json
// @Param id path string true "Book ID"
// @Success 200 {object} model.Book
// @Router /api/books/{id} [delete]
func (c *BookController) Delete(ctx *fiber.Ctx) error {
	id := ctx.Params("id")
	if id == "" {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusBadRequest, "book id cannot be empty", nil)
	}

	err := c.BookService.Delete(id)
	if err != nil {
		return utils.ErrorResponse(ctx, c.Logger, fiber.StatusInternalServerError, "failed to delete book", err)
	}

	return utils.SuccessPlainResponse(ctx, c.Logger, fiber.StatusOK, nil)
}